# apiLayer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api UMbook",
        "description": "Aqui encontra-se a documentação de toda a api de dados da UMbook, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota.\n Muitas das rotas encontram-se protegidas, não se esqueça de realizar o login na aba dos utilizadores com as suas credenciais para ter acesso às rotas que lhe são permitido aceder. Receberá um token de acesso e terá que o copiar e colar no botão 'Authorize' no canto superior direito."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3049",
    "basePath": "/api",
...
    "paths": {
...
        "/eventos": {
            "get": {
                "tags":[
                    "eventos"
                ],
                "summary": "Retorna todos os eventos registados.",
                "description": "Retorna uma lista com todos os eventos presentes.",
                "responses": {
                    "200": {
                        "description": "Retorna toda a informação relativa aos eventos que possuem uma determinada parceria.",
                        "schema": {
                            "type": "array",
                            "items":{
                                "properties": {
                                    "info": {
                                        "$ref": "#/definitions/Evento"
                                    },
                                    "participantes":{
                                        "type":"array",
                                        "items": {
                                            "type":"object",
                                            "properties": {
                                                "idUtilizador":{
                                                    "type":"string"
                                                },
                                                "nome":{
                                                    "type":"string"
                                                }
                                            }
                                        }
                                    },
                                    "parcerias":{
                                        "type": "array",
                                        "items": {
                                            "type":"object",
                                            "properties": {
                                                "idCurso":{
                                                    "type":"string"
                                                },
                                                "nome":{
                                                    "type":"string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                      },
                      "401":{
                        "description": "Falta-lhe o token de acesso ou o token não é válido."
                      }
                  },
                  "security": [
                    {
                      "api_key": []
                    }
                  ]
            },
            "post":{
                "tags":[
                    "eventos"
                ],
                "summary": "Inserir um novo evento.",
                "description": "Para inserir um novo evento, é necessário ser Admin.",
                "parameters": [
                    {
                      "name": "nome",
                      "in": "formData",
                      "description": "Nome do evento",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Informação adicional do evento",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dataInicio",
                        "in": "formData",
                        "description": "Data e hora a que o evento começa",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "dataFim",
                        "in": "formData",
                        "description": "Data e hora a que o evento acaba",
                        "required": true,
                        "type": "string"
                    },
                    {
                      "name": "parcerias",
                      "in": "formData",
                      "description": "As parcerias do Evento",
                      "required": true,
                      "type": "array",
                      "items":{
                        "type": "string"
                      }
                  }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id do evento criado",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    },
                    "403":{
                      "description": "Não possuí nível de acesso suficiente."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
        "/eventos/{eventoId}": {
            "get": {
              "tags": [
                "eventos"
              ],
              "summary": "Dado um identificador de um evento, devolve toda a informação relativa a esse evento.",
              "description": "Devolve toda a informação relativa a um evento.",
              "operationId": "getEventoById",
              "parameters": [
                {
                  "name": "eventoId",
                  "in": "path",
                  "description": "ID do evento a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna toda a informação relativa a um evento.",
                    "schema": {
                        "type": "object",
                        "properties": {
                            "info": {
                                "$ref": "#/definitions/Evento"
                            },
                            "participantes":{
                                "type":"array",
                                "items": {
                                    "type":"object",
                                    "properties": {
                                        "idUtilizador":{
                                            "type":"string"
                                        },
                                        "nome":{
                                            "type":"string"
                                        }
                                    }
                                }
                            },
                            "parcerias":{
                                "type": "array",
                                "items": {
                                    "type":"object",
                                    "properties": {
                                        "idCurso":{
                                            "type":"string"
                                        },
                                        "nome":{
                                            "type":"string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "delete": {
              "tags": [
                "eventos"
              ],
              "summary": "Apaga um determinado evento.",
              "description": "Apenas um utilizador do tipo Admin, pode apagar Eventos.",
              "operationId": "deleteEvento",
              "parameters": [
                {
                  "name": "eventoId",
                  "in": "path",
                  "description": "ID do evento.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "O evento foi apagado com sucesso."
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            }
        },
...
        "/publicacoes/{publicacaoId}": {
            "get": {
              "tags": [
                "publicacoes"
              ],
              "summary": "Dado um identificador de uma publicação, devolve toda a informação dessa publicação.",
              "description": "Devolve todas as informações relativas a uma publicação.",
              "operationId": "getPublicacaoById",
              "parameters": [
                {
                  "name": "publicacaoId",
                  "in": "path",
                  "description": "ID da publicação a retornar.",
                  "required": true,
                  "type": "string"
                }
              ],
              "responses": {
                "200": {
                    "description": "Retorna toda a informação relativa a uma publicação.",
                    "schema": {
                        "$ref": "#/definitions/Publicacao"
                    } 
                },
                "401":{
                  "description": "Falta-lhe o token de acesso ou o token não é válido."
                }
              },
              "security": [
                {
                  "api_key": []
                }
              ]
            },
            "put":{
                "tags":[
                    "publicacoes"
                ],
                "summary": "Edita uma publicação com um determinado identificador.",
                "description": "Editar publicação.",
                "parameters": [
                    {
                      "name": "publicacaoId",
                      "in": "path",
                      "description": "Identificador da publicação",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Novo conteúdo da publicação",
                        "required": true,
                        "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id da publicação editada",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            },
            "delete":{
                "tags":[
                    "publicacoes"
                ],
                "summary": "Apaga uma publicação com um determinado identificador.",
                "description": "Apagar publicação.",
                "parameters": [
                    {
                      "name": "publicacaoId",
                      "in": "path",
                      "description": "Identificador da publicação",
                      "required": true,
                      "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna 200 se a publicação foi apagada"
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/publicacoes/search/{titulo}":{
          "get": {
            "tags": [
              "publicacoes"
            ],
            "summary": "Dado um determinado título, devolve todas as publicações que contém esse título.",
            "description": "Devolve todas publicações que contém um título.",
            "operationId": "getPublicacaoById",
            "parameters": [
              {
                "name": "titulo",
                "in": "path",
                "description": "Titulo a encontrar.",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                  "description": "Retorna toda as publicações encontradas.",
                  "schema":{
                    "type": "array",
                    "items":{
                        "properties":{
                          "schema": {
                            "$ref": "#/definitions/Publicacao"
                          }
                        }
                    }
                  }
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
        },
...
        "/publicacoes":{
            "post":{
                "tags":[
                    "publicacoes"
                ],
                "summary": "Inserir uma nova publicação.",
                "description": "Adiciona uma nova publicação à base de dados.",
                "parameters": [
                    {
                      "name": "idGrupo",
                      "in": "formData",
                      "description": "Id do grupo em que a publicação vai ser publicada",
                      "required": true,
                      "type": "string"
                    },
                    {
                        "name": "conteudo",
                        "in": "formData",
                        "description": "Conteudo da publicação",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "idUtilizador",
                        "in": "formData",
                        "description": "Id do utilizador que a publicou",
                        "required": true,
                        "type": "string"
                    }
                  ],
                "responses": {
                    "200": {
                        "description": "Retorna o id da publicação criada",
                        "schema": {
                            "properties": {
                                "id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401":{
                      "description": "Falta-lhe o token de acesso ou o token não é válido."
                    }
                },
                "security": [
                  {
                    "api_key": []
                  }
                ]
            }
        },
...
        "/utilizadores":{
          "get":{
            "tags":[
              "utilizadores"
          ],
          "summary": "Retorna todos os utilizadores da aplicação.",
          "description": "Para tal, é necessário o utilizador que faz o pedido ser Admin.",
          "responses": {
            "200": {
              "description": "Lista de Utilizadores.",
              "schema":{
                "type":"array",
                "items": {
                  "properties": {
                    "schema": {
                      "$ref": "#/definitions/Utilizador"
                    }  
                  }
                }
              }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            },
            "403":{
              "description": "Não possuí nível de acesso suficiente."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
          },
          "post": {
            "tags":[
              "utilizadores"
            ],
            "summary": "Insere um novo utilizador.",
            "description": "O utilizador é adicionado á base de dados.",
            "parameters": [
              {
                "name": "id",
                "in": "formData",
                "description": "Email do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "nome",
                "in": "formData",
                "description": "Nome do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroAluno",
                "in": "formData",
                "description": "Número de aluno",
                "required": true,
                "type": "string"
              },
              {
                "name": "idCurso",
                "in": "formData",
                "description": "Identificador do curso a que pertence",
                "required": true,
                "type": "string"
              },
              {
                "name": "idAno",
                "in": "formData",
                "description": "Identificador do ano que frequenta",
                "required": true,
                "type": "string"
              },
              {
                "name": "dataNascimento",
                "in": "formData",
                "description": "Data de Nascimento",
                "required": true,
                "type": "string"
              },
              {
                "name": "sexo",
                "in": "formData",
                "description": "Género do utilizador",
                "enum": [ "Masculino", "Feminino", "Outro"],
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroTelemovel",
                "in": "formData",
                "description": "Número de telemóvel",
                "required": true,
                "type": "string"
              },
              {
                "name": "password",
                "in": "formData",
                "description": "Password",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "O utilizador foi inserido com sucesso."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
        },
        "/utilizadores/admin":{
          "post": {
            "tags":[
              "utilizadores"
            ],
            "summary": "Insere um novo utilizador de um determinado tipo.",
            "description": "Dado que se pode escolher o tipo de utilizador, é necessário ser Admin.",
            "parameters": [
              {
                "name": "tipo",
                "in": "formData",
                "description": "Tipo do utilizador",
                "enum": [ "Admin", "Responsavel", "Aluno"],
                "required": true,
                "type": "string"
              },
              {
                "name": "id",
                "in": "formData",
                "description": "Email do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "nome",
                "in": "formData",
                "description": "Nome do utilizador",
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroAluno",
                "in": "formData",
                "description": "Número de aluno",
                "required": true,
                "type": "string"
              },
              {
                "name": "idCurso",
                "in": "formData",
                "description": "Identificador do curso a que pertence",
                "required": true,
                "type": "string"
              },
              {
                "name": "idAno",
                "in": "formData",
                "description": "Identificador do ano que frequenta",
                "required": true,
                "type": "string"
              },
              {
                "name": "dataNascimento",
                "in": "formData",
                "description": "Data de Nascimento",
                "required": true,
                "type": "string"
              },
              {
                "name": "sexo",
                "in": "formData",
                "description": "Género do utilizador",
                "enum": [ "Masculino", "Feminino", "Outro"],
                "required": true,
                "type": "string"
              },
              {
                "name": "numeroTelemovel",
                "in": "formData",
                "description": "Número de telemóvel",
                "required": true,
                "type": "string"
              },
              {
                "name": "password",
                "in": "formData",
                "description": "Password",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "O utilizador foi inserido com sucesso."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              },
              "403":{
                "description": "Não possuí nível de acesso suficiente."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
        },
        "/utilizadores/{utilizadorId}/pedidosamizade":{
          "post":{
            "tags":[
              "utilizadores"
            ],
            "summary": "Adiciona um pedido de amizade.",
            "description": "Dados dois identificadores de utilizadores, o pedido de amizade é realizado.",
            "parameters": [
              {
                "name": "utilizadorId",
                "in": "path",
                "description": "Id do utilizador que faz o pedido",
                "required": true,
                "type": "string"
              },
              {
                "name": "idUtilizador2",
                "in": "formData",
                "description": "Id do utilizador que recebe o pedido",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "O pedido foi efetuado com sucesso."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]
          }
        },
...
        "/utilizadores/pedidosamizade/{idUtilizador1}/{idUtilizador2}":{
          "delete":{
            "tags": [
              "utilizadores"
            ],
            "summary": "Apaga/Rejeita um pedido de amizade.",
            "description": "Dados os dois utilizadores do pedido de amizade, o pedido é removido",
            "operationId": "deletePedidoAmizade",
            "parameters": [
              {
                "name": "idUtilizador1",
                "in": "path",
                "description": "ID do utilizador que fez o pedido.",
                "required": true,
                "type": "string"
              },
              {
                "name": "idUtilizador2",
                "in": "path",
                "description": "ID do utilizador que recebeu o pedido.",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
              "200": {
                "description": "O pedido de amizade foi eliminado/rejeitado com sucesso."
              },
              "401":{
                "description": "Falta-lhe o token de acesso ou o token não é válido."
              }
            },
            "security": [
              {
                "api_key": []
              }
            ]  
          }
        },
...
    },
      "securityDefinitions":{
        "api_key":{
          "type": "apiKey",
          "in": "query",
          "name": "token"
        }
      },
      "security": [{
         "api_key": []
      }
    ],
    "definitions": {
        "Utilizador": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "nomeUtilizador": {
                "type": "string"
              },
              "numeroAluno": {
                "type": "string"
              },
              "dataNascimento": {
                "type": "string"
              },
              "numeroTelemovel": {
                "type": "string"
              },
              "sexo": {
                "type": "string"
              }
            }
        },
...
        "Evento":{
            "type":"object",
            "properties": {
                "id":{
                    "type":"string"
                },
                "nome":{
                    "type":"string"
                },
                "conteudo":{
                    "type":"string"
                },
                "dataInicio":{
                    "type":"string"
                },
                "dataFim":{
                    "type":"string"
                }
            }
        },
...
    }   
}